requirements.txt
langgraph==0.2.60
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-google-genai==2.0.8
chromadb==0.5.23
sentence-transformers==3.3.1
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
streamlit==1.41.1
pandas==2.2.3
python-dotenv==1.0.1
httpx==0.28.1
google-generativeai==0.8.4
